back-end/backendClient2Database/requirements.txt
paho-mqtt==1.6.1
pymysql==1.1.0
cryptography==3.4.8